Allow copy and delete of ant default exclude files
Currently, it is not possible to copy or delete files
with the following names when using FileCopyOperation
and FileDeleteOperation:

CVS
.cvsignore
SCCS
vssver.scc
.svn
.DS_Store
.git
.gitattributes
.gitignore
.gitmodules
.hg
.hgignore
.hgsub
.hgsubstate
.hgtags
.bzr
.bzrignore

Since the glob Ant pattern that FilePath::list and
FilePath::copyRecursiveTo uses cannot pick up the files. This
change allows disabling the default exclude list when copying
transforming and deleting files. To allow e.g. deleting a
.gitignore file.

The Ant default exclude files are listed here:
https://ant.apache.org/manual/dirtasks.html#defaultexcludes